Refactored method signature inspections
