+++ resharper/test/data/daemon/Stages/Analysis/InitializeOnLoadMethodAttributeSignature.cs.gold
@@ -22,33 +22,9 @@ public class TestInitializeOnLoadMethod
     private static void JustRight2() { }
 }
 
-public class TestRuntimeInitializeOnLoadMethod
-{
-    [RuntimeInitializeOnLoadMethod]
-    public void |NotStatic|(4)() { }
-
-    [RuntimeInitializeOnLoadMethod]
-    public static |int|(5) WrongReturnType() { }
-
-    [RuntimeInitializeOnLoadMethod]
-    public static void WrongReturnType(|int arg1, int arg2|(6)) { }
-
-    [RuntimeInitializeOnLoadMethod]
-    public static void WrongTypeParameters|<T, T2, T3>|(7)() { }
-
-    [RuntimeInitializeOnLoadMethod]
-    public static void JustRight() { }
-
-    [RuntimeInitializeOnLoadMethod]
-    private static void JustRight2() { }
-}
 
 ---------------------------------------------------------
 (0): ReSharper Warning: Missing static modifier
 (1): ReSharper Warning: Incorrect return type. Expected 'void'
 (2): ReSharper Warning: Incorrect method parameters. Expected '()'
 (3): ReSharper Warning: Incorrect type parameters
-(4): ReSharper Warning: Missing static modifier
-(5): ReSharper Warning: Incorrect return type. Expected 'void'
-(6): ReSharper Warning: Incorrect method parameters. Expected '()'
-(7): ReSharper Warning: Incorrect type parameters
